// src/types.ts
export type JsonValue = JsonPrimitive | JsonObject | JsonValue[];
export interface JsonObject {
  [key: string]: JsonValue;
}
// ...
export type AgentIdentityCode = string;

export interface SkillDescription {
  skillId: string;
  skillName: string;
  skillDescription: string;
  tags: string[];
  examples?: string[] | undefined;
  inputTypes: string[];
  outputTypes: string[];
  dependencies?: JsonObject[] | undefined;
}

export interface AccessMethod {
  type: "url" | "ip" | "fqdn" | string;
  address: string;
  metadata?: JsonObject | undefined;
}

export interface AuthenticationDescription {
  type: "none" | "x509" | "bearer" | "custom" | string;
  metadata?: JsonObject | undefined;
}

export interface AgentDescription {
  agentId: AgentIdentityCode;
  name: string;
  alias?: string | undefined;
  version: string;
  description: string;
  iconAddress?: string | undefined;
  provider: string | JsonObject;
  accessAddress?: string | undefined;
  accessMethod?: AccessMethod[] | undefined;
  servingArea?: JsonObject | undefined;
  authentication?: AuthenticationDescription | undefined;
  capabilities: JsonObject;
  defaultInputTypes: string[];
  defaultOutputTypes: string[];
  skills: SkillDescription[];
  discoverable?: boolean | undefined;
  available?: boolean | undefined;
  metadata?: JsonObject | undefined;
}
// ...
export interface ToolDescriptor {
  toolId: string;
  toolName: string;
  toolDescription: string;
  toolVersion: string;
  toolInputParam: JsonObject;
  toolOutputParam: JsonObject;
  metadata?: JsonObject | undefined;
}